Move the Fhir library extensions to another class like it was before
