+++ Cql/Cql.Packaging/LibraryExtensions.cs
@@ -1,34 +1,126 @@
-﻿using Hl7.Cql.Compiler;
-using Hl7.Cql.Runtime;
+﻿using Hl7.Cql.Graph;
+using Hl7.Fhir.Model;
 using JetBrains.Annotations;
-using Microsoft.Extensions.Logging;
+using Hl7.Cql.Fhir;
+using FhirLibrary = Hl7.Fhir.Model.Library;
+using System.Runtime.Loader;
 
 namespace Hl7.Cql.Packaging
 {
-    /// <summary>
-    /// Helpers for ELM Libraries
-    /// </summary>
-    public static class LibraryExtensions
+    internal static class LibraryExtensions
     {
-        /// <summary>
-        /// Evaluate the given expression for the given library and context
-        /// </summary>
-        /// <param name="library">the library</param>
-        /// <param name="expression">the expression to evaluate</param>
-        /// <param name="context">the context of the expression evaluation</param>
-        /// <param name="logFactory">logger</param>
-        /// <returns>the result of the expression</returns>
         [UsedImplicitly]
-        public static object? EvaluateExpression(
-            this Elm.Library library, 
-            Elm.Expression expression, 
-            CqlContext context,
-            ILoggerFactory logFactory)
+        public static DirectedGraph GetDependencies(
+            this FhirLibrary library,
+            Func<string, FhirLibrary> getLibrary,
+            string libraryProperty = "ElmLibrary")
         {
-            var expressionBuilder = new ExpressionBuilderFactory(logFactory).ExpressionBuilderService;
-            var lambda = expressionBuilder.Lambda(library, expression);
-            var func = lambda.Compile();
-            return func.DynamicInvoke(context);
+            var graph = new DirectedGraph();
+            graph.AddStartNode();
+            Visit(null);
+            return graph;
+
+            void Visit(DirectedGraphNode? from)
+            {
+                if (!graph.Nodes.TryGetValue(library.Id, out var thisNode))
+                {
+                    thisNode = new DirectedGraphNode
+                    {
+                        NodeId = library.Id,
+                        Properties = new Dictionary<string, object>
+                        {
+                            { libraryProperty, library }
+                        }
+                    };
+                    graph.AddNode(thisNode);
+                    graph.AddEdge((from, thisNode));
+                }
+                var dependencies = library.RelatedArtifact
+                    .Where(ra => ra.Type == RelatedArtifact.RelatedArtifactType.DependsOn)
+                    .ToArray();
+
+                if (dependencies.Length == 0
+                    && !graph.GetForwardNodeIds(thisNode.NodeId).Contains(DirectedGraphNode.EndId))
+                {
+                    graph.AddEdge((thisNode, null));
+                }
+                else
+                {
+                    foreach (var dependency in dependencies)
+                    {
+                        var relatedArtifact = getLibrary(dependency.Resource)
+                            ?? throw new InvalidOperationException($"Could not find related library {dependency.ElementId}");
+                        Visit(thisNode);
+                    }
+                }
+            }
+        }
+
+        [UsedImplicitly]
+        public static IEnumerable<FhirLibrary> AllLibraries(this DirectedGraph graph, string libraryProperty = "Library")
+        {
+            IEnumerable<FhirLibrary> GetLibraries()
+            {
+                foreach (var node in graph.Nodes)
+                {
+                    if (node.Value.Properties != null)
+                    {
+                        if (node.Value.Properties.TryGetValue(libraryProperty, out object? libObject)
+                            && libObject is FhirLibrary library)
+                        {
+                            yield return library;
+                        }
+                    }
+                }
+            }
+            var libs = GetLibraries()
+                .Distinct()
+                .ToArray();
+            return libs;
+        }
+
+        [UsedImplicitly]
+        public static DirectedGraph GetDependencies(
+            this FhirLibrary library, 
+            DirectoryInfo directory) =>
+            library.GetDependencies(id =>
+            {
+                var file = new FileInfo(Path.Combine(directory.FullName, $"{id}.json"));
+                using var fs = file.OpenRead();
+                return fs.ParseFhir<FhirLibrary>();
+            });
+
+        [UsedImplicitly]
+        public static AssemblyLoadContext LoadAssemblies(
+            this IEnumerable<FhirLibrary> libraries,
+            AssemblyLoadContext asmContext)
+        {
+            foreach (var library in libraries)
+            {
+                var dll = library.Content.SingleOrDefault(att => att.ContentType == "application/octet-stream");
+                if (dll != null)
+                {
+                    using var ms = new MemoryStream(dll.Data);
+                    _ = asmContext.LoadFromStream(ms);
+                }
+
+            }
+            return asmContext;
+        }
+
+        [UsedImplicitly]
+        public static AssemblyLoadContext LoadAssembles(this IEnumerable<Binary> binaries,
+            AssemblyLoadContext asmContext)
+        {
+            foreach (var binary in binaries)
+            {
+                if (binary.ContentType == "application/octet-stream")
+                {
+                    using var ms = new MemoryStream(binary.Data);
+                    _ = asmContext.LoadFromStream(ms);
+                }
+            }
+            return asmContext;
         }
     }
 }
